LayoutTests/storage/indexeddb/deleteObjectStore-name-argument-required-expected.txt

Test IndexedDB deleteObjectStore required argument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
PASS db.deleteObjectStore(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
